+++ src/mpmToolbox/gui/msmTree/MsmTree.java
@@ -13,6 +13,8 @@ import nu.xom.Attribute;
 import nu.xom.Element;
 import nu.xom.Node;
 
+import javax.swing.event.TreeSelectionEvent;
+import javax.swing.event.TreeSelectionListener;
 import javax.swing.tree.TreePath;
 import javax.swing.tree.TreeSelectionModel;
 import java.util.ArrayList;
@@ -22,7 +24,7 @@ import java.util.Enumeration;
  * A custom WebAsncTree for MSM data.
  * @author Axel Berndt
  */
-public class MsmTree extends WebExTree<MsmTreeNode> /*implements MouseListener, TreeSelectionListener*/ {
+public class MsmTree extends WebExTree<MsmTreeNode> implements TreeSelectionListener/*, MouseListener*/ {
     @NotNull private final ProjectPane projectPane;                         // a link to the parent project pane to access its data, midi player etc.
     private WebDockableFrame dockableFrame = null;                          // a WebDockableFrame instance that displays this MSM tree, to be used in class ProjectPane
 
@@ -42,7 +44,7 @@ public class MsmTree extends WebExTree<MsmTreeNode> /*implements MouseListener, 
 //        msmTree.setCellEditor(new MsmTreeCellEditor());
 //        msmTree.setStyleId(StyleId.treeTransparent);
 
-//        this.addTreeSelectionListener(this);
+        this.addTreeSelectionListener(this);
 //        this.addMouseListener(this);
     }
 
@@ -53,11 +55,23 @@ public class MsmTree extends WebExTree<MsmTreeNode> /*implements MouseListener, 
 //    @Override
 //    public void mouseClicked(MouseEvent mouseEvent) {
 //        MsmTreeNode n = this.getNodeForLocation(mouseEvent.getX(), mouseEvent.getY());
-//        if (n != null) {
-//            n.play();
-//            System.out.println(n.getUserObject().toXML());
-//        }
+//        if (n != null)
+//            n.play(this.projectPane.getParentMpmToolbox().getMidiPlayerForSingleNotes());
 //    }
+
+    /**
+     * when a tree node is clicked, do this
+     * @param treeSelectionEvent
+     */
+    @Override
+    public void valueChanged(TreeSelectionEvent treeSelectionEvent) {
+        TreePath path = treeSelectionEvent.getNewLeadSelectionPath();
+        if (path == null)
+            return;
+
+        MsmTreeNode n = this.getNodeForPath(path);
+        n.play(this.projectPane.getParentMpmToolbox().getMidiPlayerForSingleNotes());   // the node might be a node and should play its note via MIDI when selected
+    }
 
     /**
      * a getter to access the project pane that this tree belongs to